Force enum value always output as int type
 - The compiler will complain `warning: passing
   'dmtcp::WorkerState::eWorkerState' chooses 'int' over
   'long unsigned int'` in Android build...
[Patch Supplied by Kito Cheng.]

+++ dmtcp/src/dmtcpmessagetypes.cpp
@@ -121,7 +121,7 @@ dmtcp::ostream& dmtcp::operator << ( dmtcp::ostream& o, const dmtcp::WorkerState
       OSHIFTPRINTF ( REFILLED )
     default:
       JASSERT ( false ) .Text ( "Invalid WorkerState" );
-      o << s.value();
+      o << (int)s.value();
   }
   return o;
 }
